backend/requirements.txt
apipkg==1.5
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
coverage==4.5.1
Django==2.1
django-appconf==1.0.2
django-imagekit==4.0.2
djangorestframework==3.8.2
execnet==1.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==4.3.0
pilkit==2.0
Pillow==5.2.0
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.4
pylint==2.1.1
pylint-django==2.0.1
pylint-plugin-utils==0.4
pytest==3.7.2
pytest-cov==2.5.1
pytest-django==3.4.2
pytest-forked==0.2
pytest-xdist==1.23.0
pytz==2018.5
six==1.11.0
requests==2.19.1
typed-ast==1.1.0
wrapt==1.10.11
yapf==0.22.0